+++ sigmoid.h
@@ -1,2 +1,2 @@
-float sigmoid(float x);
-float sigmoid_derivative(float x);
+double sigmoid(double x);
+double sigmoid_derivative(double x);
